pyproject.toml
[tool.poetry]
name = "RedSquirrel_0.0"
version = "0.0.1"
description = "Batch Processing of multiple assignments using streamlit and neo4j as DB, adding on to AgenticSystems"
authors = ["kkrishnan"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = "^0.11.0"
crewai_tools = "*"
duckduckgo-search = "^4.4.2"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
langchain-openai = "^0.0.5"
langchain_openai = "*"
langchain_core = "*"
langchain = "*"  # Includes everything in LangChain
langchain_community = "*"
faiss-cpu = "^1.7.1"  
streamlit = "^1.38.0"
langchain-groq = "*"
langchain_groq = "*"
langchain_experimental = "*"
neo4j = "*"


[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"